refactor: 궁성 내부에서 특정 방향으로 이동할 수 있는지 여부 반환하도록 변경
